apps/sandboxed-react-agent-authz/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
PyYAML>=6.0.2,<7
